Separate aliveness types to own file
